move packages to improve modularity and organization
