+++ application/src/main/java/org/mandarin/booking/adapter/security/CustomAuthenticationProvider.java
@@ -3,7 +3,7 @@ package org.mandarin.booking.adapter.security;
 import lombok.RequiredArgsConstructor;
 import org.mandarin.booking.AuthException;
 import org.mandarin.booking.adapter.CustomMemberAuthenticationToken;
-import org.mandarin.booking.app.persist.MemberQueryRepository;
+import org.mandarin.booking.app.member.MemberQueryRepository;
 import org.mandarin.booking.domain.member.Member;
 import org.springframework.security.authentication.AuthenticationProvider;
 import org.springframework.security.core.Authentication;
