+++ application/src/main/java/org/mandarin/booking/app/member/MemberQueryRepository.java
@@ -1,4 +1,4 @@
-package org.mandarin.booking.app.persist;
+package org.mandarin.booking.app.member;
 
 import java.util.Optional;
 import lombok.RequiredArgsConstructor;
