+++ application/src/main/java/org/mandarin/booking/app/show/ShowQueryRepository.java
@@ -1,4 +1,4 @@
-package org.mandarin.booking.app.persist;
+package org.mandarin.booking.app.show;
 
 
 import static org.mandarin.booking.domain.show.QShowSchedule.showSchedule;
