+++ application/src/test/java/org/mandarin/booking/IntegrationTestUtils.java
@@ -12,9 +12,9 @@ import java.util.Collection;
 import java.util.List;
 import java.util.UUID;
 import org.mandarin.booking.adapter.TokenUtils;
-import org.mandarin.booking.app.persist.HallCommandRepository;
-import org.mandarin.booking.app.persist.MemberCommandRepository;
-import org.mandarin.booking.app.persist.ShowCommandRepository;
+import org.mandarin.booking.app.member.MemberCommandRepository;
+import org.mandarin.booking.app.show.ShowCommandRepository;
+import org.mandarin.booking.app.venue.HallCommandRepository;
 import org.mandarin.booking.domain.member.Member;
 import org.mandarin.booking.domain.member.Member.MemberCreateCommand;
 import org.mandarin.booking.domain.member.SecurePasswordEncoder;
