+++ application/src/test/java/org/mandarin/booking/webapi/auth/login/POST_specs.java
@@ -20,7 +20,7 @@ import org.junit.jupiter.params.provider.MethodSource;
 import org.mandarin.booking.IntegrationTest;
 import org.mandarin.booking.IntegrationTestUtils;
 import org.mandarin.booking.TokenHolder;
-import org.mandarin.booking.app.persist.MemberQueryRepository;
+import org.mandarin.booking.app.member.MemberQueryRepository;
 import org.mandarin.booking.domain.member.AuthRequest;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Value;
